перенесла в модуль renderComments
